+++ NEWS.md
@@ -30,6 +30,16 @@ to vector, matrix, or array depending on the dimensions of the table. (#528)
 * `install_cmdstan()` now automatically installs the Linux ARM CmdStan when
 Linux distributions running on ARM CPUs are detected. (#531)
 
+* The `dir` argument for `write_stan_file()` can now be set with a global
+option. (#537)
+
+* Improved processing of named lists supplied to the `data` argument to JSON
+data files: checking whether the list includes all required elements/Stan
+variables; improved differentiating arrays/vectors of length 1 and scalars
+when generating JSON data files; generating floating point numbers with
+decimal points to fix issue with parsing large numbers. (#538)
+
+
 # cmdstanr 0.4.0
 
 ### Bug fixes